[KYUUBI #2028] Hive Backend Engine - Events support
