+++ externals/kyuubi-hive-sql-engine/src/main/scala/org/apache/kyuubi/engine/hive/HiveSQLEngine.scala
@@ -17,6 +17,8 @@
 
 package org.apache.kyuubi.engine.hive
 
+import java.net.InetAddress
+
 import scala.util.control.NonFatal
 
 import org.apache.hadoop.hive.conf.HiveConf
@@ -24,13 +26,14 @@ import org.apache.hadoop.hive.conf.HiveConf.ConfVars
 
 import org.apache.kyuubi.{Logging, Utils}
 import org.apache.kyuubi.config.KyuubiConf
-import org.apache.kyuubi.config.KyuubiConf.ENGINE_EVENT_LOGGERS
-import org.apache.kyuubi.engine.hive.event.{HiveEngineEvent, HiveEventLoggingService}
-import org.apache.kyuubi.events.EventLogging
+import org.apache.kyuubi.config.KyuubiConf.{ENGINE_EVENT_JSON_LOG_PATH, ENGINE_EVENT_LOGGERS}
+import org.apache.kyuubi.engine.hive.events.HiveEngineEvent
+import org.apache.kyuubi.engine.hive.events.handler.HiveJsonLoggingEventHandler
+import org.apache.kyuubi.events.{EventBus, EventLoggerType, KyuubiEvent}
 import org.apache.kyuubi.ha.HighAvailabilityConf.HA_ZK_CONN_RETRY_POLICY
 import org.apache.kyuubi.ha.client.RetryPolicies
 import org.apache.kyuubi.service.{AbstractBackendService, AbstractFrontendService, Serverable}
-import org.apache.kyuubi.util.SignalRegister
+import org.apache.kyuubi.util.{KyuubiHadoopUtils, SignalRegister}
 
 class HiveSQLEngine extends Serverable("HiveSQLEngine") {
   override val backendService: AbstractBackendService = new HiveBackendService(this)
@@ -54,13 +57,11 @@ object HiveSQLEngine extends Logging {
   kyuubiConf.set(ENGINE_EVENT_LOGGERS.key, "JSON")
 
   def startEngine(): HiveSQLEngine = {
-    val eventLogging = new HiveEventLoggingService()
     try {
-      eventLogging.initialize(kyuubiConf)
-      eventLogging.start()
+      initLoggerEventHandler(kyuubiConf)
     } catch {
       case NonFatal(e) =>
-        warn(s"Failed to initialize EventLoggingService: ${e.getMessage}", e)
+        warn(s"Failed to initialize Logger EventHandler: ${e.getMessage}", e)
     }
     kyuubiConf.setIfMissing(KyuubiConf.FRONTEND_THRIFT_BINARY_BIND_PORT, 0)
     kyuubiConf.setIfMissing(HA_ZK_CONN_RETRY_POLICY, RetryPolicies.N_TIME.toString)
@@ -86,14 +87,31 @@ object HiveSQLEngine extends Logging {
     val engine = new HiveSQLEngine()
     info(s"Starting ${engine.getName}")
     engine.initialize(kyuubiConf)
-    EventLogging.onEvent(HiveEngineEvent(engine))
+    EventBus.post(HiveEngineEvent(engine))
     engine.start()
     val event = HiveEngineEvent(engine)
     info(event)
-    EventLogging.onEvent(event)
+    EventBus.post(event)
     Utils.addShutdownHook(() => engine.stop())
     currentEngine = Some(engine)
     engine
+  }
+
+  private def initLoggerEventHandler(conf: KyuubiConf): Unit = {
+    val hadoopConf = KyuubiHadoopUtils.newHadoopConf(conf)
+    conf.get(ENGINE_EVENT_LOGGERS).map(EventLoggerType.withName).foreach {
+      case EventLoggerType.JSON =>
+        val hostName = InetAddress.getLocalHost.getCanonicalHostName
+        val handler = HiveJsonLoggingEventHandler(
+          s"HIVE-$hostName",
+          ENGINE_EVENT_JSON_LOG_PATH,
+          hadoopConf,
+          conf)
+        EventBus.register[KyuubiEvent](handler)
+      case logger =>
+        throw new IllegalArgumentException(s"Unrecognized event logger: $logger")
+
+    }
   }
 
   def main(args: Array[String]): Unit = {
@@ -105,7 +123,7 @@ object HiveSQLEngine extends Logging {
           case Some(engine) =>
             engine.stop()
             val event = HiveEngineEvent(engine).copy(diagnostic = t.getMessage)
-            EventLogging.onEvent(event)
+            EventBus.post(event)
           case _ =>
             error(s"Failed to start Hive SQL engine: ${t.getMessage}.", t)
         }
