+++ externals/kyuubi-hive-sql-engine/src/main/scala/org/apache/kyuubi/engine/hive/operation/ExecuteStatement.scala
@@ -21,7 +21,7 @@ import scala.collection.JavaConverters._
 
 import org.apache.hive.service.cli.operation.Operation
 
-import org.apache.kyuubi.engine.hive.event.HiveOperationEvent
+import org.apache.kyuubi.engine.hive.events.HiveOperationEvent
 import org.apache.kyuubi.events.EventLogging
 import org.apache.kyuubi.operation.OperationType
 import org.apache.kyuubi.session.Session
